requirements.txt
pycryptodome
boto3
pyAesCrypt
argparse
stegano

